requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0
python-dotenv==1.0.0
eventlet==0.33.3
gunicorn==21.2.0
requests==2.31.0
dnspython==2.4.2